dev-requirements.txt
-r requirements.txt
django-debug-toolbar
django-unused-media